Do not remove translator.py during code_generation.py

+++ .ci/code_generation.py
@@ -21,6 +21,8 @@ for file_path in files:
         continue
     if file_path.name == "operator.mustache":
         continue
+    if file_path.stem == "translator":
+        continue
     try:
         if file_path.is_dir():
             shutil.rmtree(file_path)
